cambios actuallizados
lo que tenia hecho pero en los archivos .kt correctos in duplicados
