+++ T3/app/src/main/res/navigation/nav_graph.xml
@@ -19,7 +19,7 @@
             app:destination="@id/homeAlumnoFragment" />
         <action
             android:id="@+id/action_loginFragment_to_homeProfesorFragment"
-            app:destination="@id/homeProfesorFragment" />
+            app:destination="@+id/homeProfesorFragment" />
     </fragment>
     <fragment
         tools:layout="@layout/fragment_registro"
